telegram_seshat/pyproject.toml
[tool.poetry]
name = "telegram-seshat"
version = "0.1.0"
description = "This app will handle the telegram side of seshat dealings.'"
authors = ["Sepehrmasihpour <sepehr.masihpour1381@gmail.com>"]
license = "Proprietary"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.0.1"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
requests = "^2.31.0"
aiohttp = "^3.9.3"
aiofiles = "^23.2.1"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

